requirements.txt
aiohttp
asyncio
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
sentence-transformers
transformers
pydantic
pydantic-settings
numpy
pytest
pytest-cov
pytest-asyncio
httpx
mypy
black
isort
python-dotenv
pre-commit
PyJWT
passlib[bcrypt]
email-validator